test(notifications): move SMS channel tests to tests/modules/notifications/ so --module scope collects them
